Quitar funciones reutilizables de AllBooksList y AllBooksListRow y ponerlas en la carpeta helperFunctions
