+++ src/components/AllBooksListRow.jsx
@@ -4,7 +4,16 @@ import checkFull from "../assets/checkFull.svg";
 import listWithCheck from "../assets/listWithCheck.svg";
 import listWithCross from "../assets/listWithCross.svg";
 import { Link } from "react-router-dom";
-import supabase from "../supabase/config";
+import {
+  markAsRead,
+  markAsUnread,
+  addToList,
+  removeFromList,
+} from "../helperFunctions/updateUserLists.js";
+import {
+  getBooksToRead,
+  getBooksRead,
+} from "../helperFunctions/getDataFromDB.js";
 
 function AllBooksListRow({
   book,
@@ -12,53 +21,7 @@ function AllBooksListRow({
   setBooksRead,
   booksToRead,
   setBooksToRead,
-  getBooksRead,
 }) {
-  async function markAsRead(book) {
-    const { data, error } = await supabase.rpc("append_book_books_read", {
-      user_id: 1,
-      new_item: Number(book.id),
-    });
-    //Usuario 1 es nuestro único usuario
-    if (error) {
-      console.log("Error: ", error);
-    }
-    getBooksRead();
-    console.log(booksRead);
-  }
-
-  /* function testFunction() {
-    if (booksRead) {
-      console.log(booksRead.includes(book.id));
-      console.log(booksRead);
-    }
-  }
-  testFunction(); */
-  function markAsUnread(book) {
-    /* const newReadBooks = [...readBooks].filter((id) => book.id !== id);
-    setBooksRead(newReadBooks);
-    book.isRead = false;
-    console.log(newReadBooks);
-    const newBooks = [...books];
-    setBooks(newBooks); */
-    console.log(book);
-  }
-
-  function addToList(book) {
-    const newBooksToRead = [...booksToRead, book.id];
-    setBooksToRead(newBooksToRead);
-    book.isInList = true;
-    console.log("newBooksToRead ", newBooksToRead);
-    console.log("booksToRead ", booksToRead);
-  }
-  function removeFromList(book) {
-    const newBooksToRead = [...booksToRead].filter((id) => book.id !== id);
-    setBooksToRead(newBooksToRead);
-    book.isInList = false;
-    console.log("newBooksToRead ", newBooksToRead);
-    console.log("booksToRead ", booksToRead);
-  }
-
   return (
     <>
       <li className="AllBooksListRow">
@@ -83,26 +46,36 @@ function AllBooksListRow({
           <img
             className="AllBooksListRow-options"
             src={checkFull}
-            onClick={() => markAsUnread(book)}
+            onClick={() => {
+              markAsUnread(book, getBooksRead, setBooksRead);
+            }}
           />
         ) : (
           <img
             className="AllBooksListRow-options"
             src={check}
-            onClick={() => markAsRead(book)}
+            onClick={() => {
+              markAsRead(book, getBooksRead, setBooksRead);
+            }}
           />
         )}
         {book.isInList ? (
           <img
             className="AllBooksListRow-options"
             src={listWithCross}
-            onClick={() => removeFromList(book)}
+            onClick={() => {
+              removeFromList(book);
+              getBooksToRead();
+            }}
           />
         ) : (
           <img
             className="AllBooksListRow-options"
             src={listWithCheck}
-            onClick={() => addToList(book)}
+            onClick={() => {
+              addToList(book);
+              getBooksToRead();
+            }}
           />
         )}
       </li>
